Add Contact list model
